Refactor test base and company integration tests
The changes removed duplicate code in the `HubSpotCompanyApiIntegrationTests` class and moved it to the `HubSpotIntegrationTestBase` class. It also removed the `IDisposable` implementation from the `HubSpotCompanyApiIntegrationTests` class and introduced it into the `HubSpotIntegrationTestBase` class. This makes the code cleaner and easier to maintain as common code is placed in the base class.
